Added buttons to view page
Added a button for each database that shows the information which has been inserted into them.

Fixed the code in routers.js and handers.js which makes the above buttons possible to work.

Created a new page for /view which holds the database buttons.

Created a page for each of the database buttons.