Version 1.0 - Added sorting and search option
Removed copy of editCity.js code in editCountry.js

Fixed a bug in editCity.js and editLanguage.js which replaced data in the database that it shouldn't. - Possible that a country has the same cityname or language that another country has. - Possible that a Language is spoken in more that 1 country. - Possible that a cityname is present in another country (This part caused a lot of late night research)

Added more styling